whisper/pyproject.toml
[project]
name = "Whisper-Trainings"
version = "0.1.0"
description = "Finetuning Whisper models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clearml[s3]>=2.0.2",
    "ctranslate2>=4.6.0",
    "datasets==3.6.0",
    "evaluate>=0.4.5",
    "huggingface-hub>=0.34.3",
    "jiwer>=4.0.0",
    "librosa>=0.11.0",
    "numba>=0.61.2",
    "openpyxl>=3.1.5",
    "peft>=0.17.0",
    "pydub>=0.25.1",
    "pynvml>=12.0.0",
    "pyyaml>=6.0.2",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "torch>=2.7.1",
    "transformers>=4.55.0",
]
